backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Sarahkh4" },
    { name = "zubair-hafeez" },
    { name = "r34g4n" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=0.3.24",
    "langchain-google-genai>=2.1.3",
    "langgraph>=0.4.0",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langgraph-swarm>=0.0.2",
    "psycopg>=3.2.6",
    "psycopg-pool>=3.2.6",
    "pydantic-settings>=2.9.1",
    "trustcall>=0.0.39",
    "tavily-python>=0.7.0",
]

[dependency-groups]
dev = [
    "gradio>=5.27.1",
    "ipykernel>=6.29.5",
    "ruff>=0.11.7",
]
